Add a FORGIVING Compiletime-flag
It is meant to be forgiving in respect to end-tags.
If the library is compiled with this flag, any missing end-tag
will be ignored, and the tree following after that event will
be utterly scrambled.

So this is quite a stupid thing to do, but unfortunately it is
necessary if someone delivers invalid XML-code.